requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
dnspython==2.4.2
python-nmap==0.7.1
colorama==0.4.6
tqdm==4.66.1
pyyaml==6.0.1
urllib3==2.0.7
selenium==4.15.2
aiohttp==3.9.1
asyncio==3.4.3
lxml==5.3.0
jinja2==3.1.2
python-whois==0.8.0
shodan==1.31.0
censys==2.2.8
builtwith==1.3.4
Wappalyzer==1.0.18
paramiko==3.3.1
pycryptodome==3.19.0
validators==0.22.0
termcolor==2.3.0
tabulate==0.9.0
python-dotenv==1.0.0
ollama==0.3.3
tweepy==4.14.0
schedule==1.2.0
